chore: Footer & Header components moved
