+++ client/src/components/index.js
@@ -4,5 +4,5 @@ export { default as Contact } from "./home/Contact.jsx";
 export { default as Courses } from "./home/Courses.jsx";
 export { default as Projects } from "./home/Projects.jsx";
 export { default as Services } from "./home/Services.jsx";
-export { default as Header } from "./layout/Header.jsx";
-export { default as Footer } from "./layout/Footer.jsx";
+export { default as Header } from "./home/Header.jsx";
+export { default as Footer } from "./common/footer/Footer.jsx";
